File restructuring has been made, review comments addressed
