+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.WeatherApp"
         tools:targetApi="31" >
         <activity
-            android:name=".models.MainActivity"
+            android:name=".view.viewactivities.MainActivity"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -25,7 +25,7 @@
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
-        <activity android:name=".models.DetailsActivity"></activity>
+        <activity android:name=".view.viewactivities.DetailsActivity"></activity>
     </application>
 
 </manifest>
