+++ app/src/main/java/com/example/weatherapp/data/repository/remoterepository/apiservices/LocationAPIService.kt
@@ -1,9 +1,9 @@
-package com.example.weatherapp.models
+package com.example.weatherapp.data.repository.remoterepository.apiservices
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
-object LocationService {
+object LocationAPIService {
 
     private const val baseUrl: String = "https://restcountries.com/"
