+++ app/src/main/java/com/example/weatherapp/data/repository/remoterepository/remoterepositoryobjects/LocationWeatherDaysData.kt
@@ -1,0 +1,8 @@
+package com.example.weatherapp.data.repository.remoterepository.remoterepositoryobjects
+
+data class LocationWeatherDaysData(
+    val temp: Double,
+    val humidity: String,
+    val description: String,
+    val icon: String,
+)
