+++ app/src/main/java/com/example/weatherapp/data/repository/remoterepository/remoterepositoryobjects/LocationWeatherRemoteData.kt
@@ -1,0 +1,6 @@
+package com.example.weatherapp.data.repository.remoterepository.remoterepositoryobjects
+
+data class LocationWeatherRemoteData(
+    val address: String?,
+    val days: List<LocationWeatherDaysData>?,
+)
