--- app/src/main/java/com/example/weatherapp/models/Location.kt
@@ -1,5 +1,0 @@
-package com.example.weatherapp.models
-
-class Location(
-    val name: LocationData
-)
