--- app/src/main/java/com/example/weatherapp/models/LocationData.kt
@@ -1,6 +1,0 @@
-package com.example.weatherapp.models
-
-data class LocationData(
-    val common: String,
-    val official: String
-)
