--- app/src/main/java/com/example/weatherapp/models/LocationWeatherData.kt
@@ -1,6 +1,0 @@
-package com.example.weatherapp.models
-
-data class LocationWeatherData(
-    val address: String?,
-    val days: List<TemperatureData>?,
-)
