--- app/src/main/java/com/example/weatherapp/models/LocationWeatherRemoteData.kt
@@ -1,6 +1,0 @@
-package com.example.weatherapp.models
-
-data class LocationWeatherRemoteData(
-    val address: String?,
-    val days: List<TemperatureData>?,
-)
