--- app/src/main/java/com/example/weatherapp/models/TemperatureData.kt
@@ -1,8 +1,0 @@
-package com.example.weatherapp.models
-
-data class TemperatureData(
-    val temp: Double,
-    val humidity: String,
-    val description: String,
-    val icon: String,
-)
