--- app/src/main/java/com/example/weatherapp/models/WeatherMainMenuUiState.kt
@@ -1,11 +1,0 @@
-package com.example.weatherapp.models
-
-data class WeatherMainMenuUiState(
-    val isPageLoaded: Boolean = false,
-    val locations: List<Location>? = listOf(),
-)
-
-data class WeatherMainMenuItemState(
-    val locationName: String,
-    val locationContinent: String
-)
