--- app/src/main/java/com/example/weatherapp/models/WeatherPost.kt
@@ -1,6 +1,0 @@
-package com.example.weatherapp.models
-
-data class WeatherPost(
-    val city: String,
-    val temperature: String,
-)
