+++ app/src/main/java/com/example/weatherapp/viewmodel/MainMenuViewModel.kt
@@ -1,12 +1,14 @@
-package com.example.weatherapp.models
+package com.example.weatherapp.viewmodel
 
 import kotlinx.coroutines.flow.MutableStateFlow
 import androidx.lifecycle.ViewModel
+import com.example.weatherapp.data.repository.LocationRepository
+import com.example.weatherapp.viewmodel.uistate.WeatherMainMenuUiState
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 
-class WeatherMainMenuViewModel : ViewModel() {
+class MainMenuViewModel : ViewModel() {
     private val _uiState = MutableStateFlow(WeatherMainMenuUiState())
     val uiState: StateFlow<WeatherMainMenuUiState> = _uiState.asStateFlow()
