+++ app/src/main/java/com/example/weatherapp/viewmodel/WeatherDetailsViewModel.kt
@@ -1,6 +1,10 @@
-package com.example.weatherapp.models
+package com.example.weatherapp.viewmodel
 
 import androidx.lifecycle.ViewModel
+import com.example.weatherapp.viewmodel.objects.LocationWeatherData
+import com.example.weatherapp.data.repository.remoterepository.remoterepositoryobjects.LocationWeatherRemoteData
+import com.example.weatherapp.data.repository.WeatherDetailsRepository
+import com.example.weatherapp.viewmodel.uistate.WeatherDetailsUiState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -11,26 +15,12 @@ import kotlinx.coroutines.withContext
 class WeatherDetailsViewModel: ViewModel() {
     private val _uiState = MutableStateFlow(WeatherDetailsUiState())
     val uiState: StateFlow<WeatherDetailsUiState> = _uiState.asStateFlow()
-
     private val locationWeatherRepository = WeatherDetailsRepository()
-
-    private fun updateLocationWeatherData(selectedLocationWeatherData: LocationWeatherData?) {
-        _uiState.update { currentState ->
-            currentState.copy(weatherInformation = selectedLocationWeatherData)
-        }
-    }
 
     fun isPageLoaded(loaded: Boolean ){
         _uiState.update { currentState ->
             currentState.copy(isPageLoaded = loaded)
         }
-    }
-
-    private fun convertRemoteDataToViewModelData(remoteDataResponse: LocationWeatherRemoteData?): LocationWeatherData{
-        return LocationWeatherData(
-            address = remoteDataResponse?.address,
-            days = remoteDataResponse?.days
-        )
     }
 
     suspend fun setTemperatureDetails(location: String) {
@@ -40,5 +30,18 @@ class WeatherDetailsViewModel: ViewModel() {
             }
         val convertedRemoteDataResponse = convertRemoteDataToViewModelData(remoteDataResponse)
         updateLocationWeatherData(convertedRemoteDataResponse)
+    }
+
+    private fun updateLocationWeatherData(selectedLocationWeatherData: LocationWeatherData?) {
+        _uiState.update { currentState ->
+            currentState.copy(weatherInformation = selectedLocationWeatherData)
+        }
+    }
+
+    private fun convertRemoteDataToViewModelData(remoteDataResponse: LocationWeatherRemoteData?): LocationWeatherData {
+        return LocationWeatherData(
+            address = remoteDataResponse?.address,
+            days = remoteDataResponse?.days
+        )
     }
 }
