+++ app/src/main/java/com/example/weatherapp/viewmodel/objects/Location.kt
@@ -1,0 +1,5 @@
+package com.example.weatherapp.viewmodel.objects
+
+class Location(
+    val name: LocationName
+)
