+++ app/src/main/java/com/example/weatherapp/viewmodel/objects/LocationName.kt
@@ -1,0 +1,6 @@
+package com.example.weatherapp.viewmodel.objects
+
+data class LocationName(
+    val common: String,
+    val official: String
+)
