+++ app/src/main/java/com/example/weatherapp/viewmodel/objects/LocationWeatherData.kt
@@ -1,0 +1,8 @@
+package com.example.weatherapp.viewmodel.objects
+
+import com.example.weatherapp.data.repository.remoterepository.remoterepositoryobjects.LocationWeatherDaysData
+
+data class LocationWeatherData(
+    val address: String?,
+    val days: List<LocationWeatherDaysData>?,
+)
