+++ app/src/main/java/com/example/weatherapp/viewmodel/uistate/WeatherMainMenuUiState.kt
@@ -1,0 +1,8 @@
+package com.example.weatherapp.viewmodel.uistate
+
+import com.example.weatherapp.viewmodel.objects.Location
+
+data class WeatherMainMenuUiState(
+    val isPageLoaded: Boolean = false,
+    val locations: List<Location>? = listOf(),
+)
